CUT duplication test cases & moved pdfconduit tests into root test directory
