--- tests/pdfconduit/__init__.py
@@ -1,4 +1,0 @@
-from tests.pdfconduit._testcase import PdfconduitTestCase
-from tests.pdfconduit.test_encrypt import EncryptionTestCase
-
-__all__ = [PdfconduitTestCase, EncryptionTestCase]
